requirements.txt
asgiref==3.7.2
colorama==0.4.6
distlib==0.3.6
Django==4.2.2
django-tailwind==3.6.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
filelock==3.12.2
inflection==0.5.1
mysqlclient==2.2.0
platformdirs==3.7.0
PyMySQL==1.1.0
PyPDF2==3.0.1
pytz==2023.3
sqlparse==0.4.4
termcolor==2.3.0
tzdata==2023.3
virtualenv==20.23.1
